requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
gigachat==0.1.20
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.0
certifi==2024.2.2
python-dotenv==1.0.0